requirements.txt
python-telegram-bot
flask
moviepy
numpy
scikit-image
python-dotenv